chore: fix bugs, build test

+++ app/screens/auth/browse/attachment/item/useAttachmentActions.ts
@@ -1,9 +1,21 @@
+/* eslint-disable react-native/split-platform-components */
 import { AttachmentType } from "../usePickAttachment"
 import { attachmentApi } from "app/services/api"
 import { useStores } from "app/models"
 import { useHelper } from "app/services/hook"
 import RNFS from "react-native-fs"
 import { useCoreService } from "app/services/coreService"
+import { Platform } from "react-native"
+import { usePermission } from "../permission"
+import Share, { ShareOptions } from "react-native-share"
+import { CameraRoll } from "@react-native-camera-roll/camera-roll"
+
+const DOWNLOAD_PATH =
+  Platform.OS === "android"
+    ? Platform.Version >= 30
+      ? RNFS.DocumentDirectoryPath
+      : RNFS.DownloadDirectoryPath
+    : RNFS.DocumentDirectoryPath
 
 export const useAttachmentActions = (
   attachment: AttachmentType,
@@ -13,13 +25,20 @@ export const useAttachmentActions = (
   const { cipherStore } = useStores()
   const { attachmentService } = useCoreService()
   const { notify, notifyApiError, translate } = useHelper()
+  const { requestStoragePermission, hasAndroidGalleryPermission } = usePermission()
 
   const onDownloadAttachment = async () => {
     if (!attachment.key) return
     try {
       setIsLoading(true)
+      const hasPermission = await requestStoragePermission()
+      if (!hasPermission) {
+        console.error("onDownloadAttachment: Permission denied!")
+        return
+      }
+
       const tempEncFile = `${RNFS.CachesDirectoryPath}/${Date.now()}enc${attachment.fileName}`
-      const filePath = `${RNFS.DocumentDirectoryPath}/${Date.now()}-${attachment.fileName}`
+      const filePath = `${DOWNLOAD_PATH}/${attachment.fileName}`
       if (await RNFS.exists(filePath)) {
         await RNFS.unlink(filePath)
       }
@@ -50,7 +69,40 @@ export const useAttachmentActions = (
         notify("error", translate("file_attachment.error.decrypt_error"))
         return
       }
-      notify("success", translate("file_attachment.download_success"))
+
+      if (getFileType(attachment.fileName) === "file") {
+        if (Platform.OS === "android") {
+          if (Platform.Version >= 30) {
+            const options: ShareOptions = {
+              title: "Save File",
+              url: `file://${filePath}`,
+              saveToFiles: true,
+            }
+
+            try {
+              await Share.open(options)
+            } catch (error) {
+              if (error.message === "User did not share") {
+                // "User canceled sharing"
+              } else {
+                console.error("Error sharing:", error)
+              }
+            }
+          }
+        } else {
+          notify("success", translate("file_attachment.download_success"))
+        }
+      } else {
+        try {
+          if (Platform.OS === "android" && !(await hasAndroidGalleryPermission())) {
+            return
+          }
+          await CameraRoll.saveAsset(filePath)
+          notify("success", translate("file_attachment.download_media_success"))
+        } catch (error) {
+          console.error("Error saving media:", error)
+        }
+      }
     } catch (error) {
       notify("error", translate("file_attachment.error.download_error"))
     } finally {
@@ -72,4 +124,20 @@ export const useAttachmentActions = (
   }
 
   return { onDownloadAttachment, onDeleteAttachment }
+}
+
+const getFileType = (fileName: string): "image" | "video" | "file" => {
+  // Define extensions for images and videos
+  const imageExtensions = ["jpg", "jpeg", "png", "gif", "bmp", "webp", "tiff", "svg", "heic"]
+  const videoExtensions = ["mp4", "mkv", "avi", "mov", "flv", "wmv", "webm", "3gp", "mpeg"]
+
+  // Extract file extension
+  const extension = fileName.split(".").pop()?.toLowerCase()
+
+  if (!extension) return "file" // No extension means unknown file
+
+  if (imageExtensions.includes(extension)) return "image"
+  if (videoExtensions.includes(extension)) return "video"
+
+  return "file" // Default case for other files
 }
